sigquitを受け付けたときにカーソルが移動する不具合を修正
