+++ includes/execution.h
@@ -6,7 +6,7 @@
 /*   By: takanoraika <takanoraika@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 18:27:12 by takanoraika       #+#    #+#             */
-/*   Updated: 2022/10/12 12:10:24 by takanoraika      ###   ########.fr       */
+/*   Updated: 2022/10/14 12:32:38 by takanoraika      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,5 +24,6 @@ typedef struct s_pipe {
 
 int		builtin_check_and_run(char **args);
 int		bin_check_and_run(char **args);
+void	exec_signal_set(void);
 
 #endif
