+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: takanoraika <takanoraika@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 11:20:30 by shima             #+#    #+#             */
-/*   Updated: 2022/10/12 14:39:22 by takanoraika      ###   ########.fr       */
+/*   Updated: 2022/10/14 12:54:29 by takanoraika      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
